validate request and request canceling

+++ hr_system/static/js/request.js
@@ -1,4 +1,3 @@
-
 function getCookie(name) {
     let cookieValue = null;
     if (document.cookie && document.cookie !== '') {
@@ -19,6 +18,8 @@ const csrftoken = getCookie('csrftoken');
 
 function reset(){
  document.getElementById('form-wrapper').reset()
+ document.getElementById('checkedList').innerHTML=''
+ document.getElementById('uncheckedList').innerHTML=''
 
 }
 
@@ -27,10 +28,8 @@ function reset(){
 
 // post
 function postRequest(userID){
-id=userID;
-    var form = document.getElementById('form-wrapper')
-
-    console.log('submitted')
+    infoList(userID);
+    var form = document.getElementById('form-wrapper');
 
     var url = `http://127.0.0.1:8000/api/requests/`
 
@@ -42,26 +41,66 @@ id=userID;
     var checked = false;
     var desc = 'No description';
 
+    var startmoment= moment(startd)
+    var endmoment= moment(endd)
 
-    fetch(url, {
-        method : 'POST',
-         headers:{'Content-type' : 'application/json',
-                 'X-CSRFToken': csrftoken
-        },
-        body : JSON.stringify({
-            'user_id': userID,
-            'start_date':startd,
-            'end_date':endd,
-            'start_hour':starth,
-            'end_hour':endh,
-            'approval_flag':approval_flag,
-            'checked': checked,
-            'description':desc
+    var starthour = moment(starth,"hh:mm")
+
+    var endhour = moment(endh,"hh:mm")
+
+    //alert(starthour.isBefore(endhour))
+
+    if(endmoment.isBefore(startmoment)){
+        alert('Your input is invalid. Please make sure your start date is before your end date!')
+        return;
+    }else if(startmoment.isSame(endmoment)){
+        if(endhour.isBefore(starthour)){
+            alert('Your input is invalid. Please make sure your start hour is before your end hour!')
+            return;
+        }
+        else{
+            //subtract from total hour
+            fetch(url, {
+                method : 'POST',
+                 headers:{'Content-type' : 'application/json',
+                         'X-CSRFToken': csrftoken
+                },
+                body : JSON.stringify({
+                    'user_id': userID,
+                    'start_date':startd,
+                    'end_date':endd,
+                    'start_hour':starth,
+                    'end_hour':endh,
+                    'approval_flag':approval_flag,
+                    'checked': checked,
+                    'description':desc
+                })
+            }).then(function(response){
+                reset();
+            })
+        }
+    }else{
+        //subtract from total days
+        fetch(url, {
+            method : 'POST',
+             headers:{'Content-type' : 'application/json',
+                     'X-CSRFToken': csrftoken
+            },
+            body : JSON.stringify({
+                'user_id': userID,
+                'start_date':startd,
+                'end_date':endd,
+                'start_hour':starth,
+                'end_hour':endh,
+                'approval_flag':approval_flag,
+                'checked': checked,
+                'description':desc
+            })
+        }).then(function(response){
+            reset();
         })
-    }).then(function(response){
-        reset();
-        //console.log(response.user_id)
-    })
+
+    }
 }
 
 
@@ -141,46 +180,62 @@ function buildList(){
 function deleteRequest(req_id){
     //console.log('clicked',req_id)
     var url = `http://127.0.0.1:8000/api/requests/${req_id}/`
-
-
+    // get the request  to be deleted
     fetch( url, {
-        method: 'DELETE',
-        headers:{'Content-type' : 'application/json',
-        'X-CSRFToken': csrftoken
-        }
-        }).then(function(response){
-            document.getElementById('checkedList').innerHTML=''
-            document.getElementById('uncheckedList').innerHTML=''
-            buildList()
+            method: 'GET',
+            headers:{'Content-type' : 'application/json',
+            'X-CSRFToken': csrftoken
+            }
+            }).then((resp)=> resp.json())
+            .then(function(response){
+            // if the request is checked than you cant cancel less then 48 hours before the startdate
+            if (response.checked){
+                // check if the canceling is being done  48 hours before the start date
+                sd = response.start_date
+                sh = response.start_hour
+                a =sd.concat(sh)
+
+                sdm = moment(a, "YYYY-MM-DDhh:mm")
+                sdm.subtract(48,"hours");
+                if(moment().isAfter(sdm)){
+                    alert("Oops!   You cannot cancel now. Too late.")
+                    return;
+                }else{
+                    fetch( url, {
+                    method: 'DELETE',
+                    headers:{'Content-type' : 'application/json',
+                    'X-CSRFToken': csrftoken
+                    }
+                    }).then(function(response){
+                        reset()
+                        buildList()
+                        })
+                }
+            }
+            else{
+                fetch( url, {
+                method: 'DELETE',
+                headers:{'Content-type' : 'application/json',
+                'X-CSRFToken': csrftoken
+                }
+                }).then(function(response){
+                    reset()
+                    buildList()
+                    })
+            }
             })
-}
 
-
-infoList(1)
-// list personal information
-function infoList(userID){
-    var url = `http://127.0.0.1:8000/api/users/${userID}`
-    fetch(url)
-    .then((resp)=> resp.json())
-    .then(function(response){
-        document.getElementById('name').innerHTML += response.first_name
-        document.getElementById('name').innerHTML += ' '
-        document.getElementById('name').innerHTML += response.last_name
-        document.getElementById('salary').innerHTML += ': '
-        document.getElementById('salary').innerHTML += response.salary
-        document.getElementById('phone').innerHTML += ': '
-        document.getElementById('phone').innerHTML += response.phone_no
-
-    })
 }
 
 
 
+
+var sd=null;
+var sh=null;
 // get one specific request
-var id=1;
-function checkRequest(){
-    var url = `http://127.0.0.1:8000/api/requests/${id}/`
-    console.log(id)
+function getRequest(user){
+    var url = `http://127.0.0.1:8000/api/requests/${user}/`
+
         fetch( url, {
             method: 'GET',
             headers:{'Content-type' : 'application/json',
@@ -188,12 +243,134 @@ function checkRequest(){
             }
             }).then((resp)=> resp.json())
             .then(function(response){
-                if(response.approval_flag){
+                sd = response.start_date
+                sh = response.start_hour
 
-                    //to be continued
-                    
-                }
                 })
     }
 
-checkRequest()
+
+
+
+// list personal information
+function infoList(userID){
+
+    var url = `http://127.0.0.1:8000/api/users/${userID}`
+    fetch(url)
+    .then((resp)=> resp.json())
+    .then(function(response){
+    document.getElementById('listt').innerHTML=`
+        <li>Date & Time: ${moment()}</li>
+        <li class="pb-3 pt-3" id="name"><strong>${response.first_name} ${response.last_name}</strong></li>
+        <li class="pb-3" id="jobid">Job position:  </li>
+        <li class="pb-3" id="salary">Salary: ${response.salary}</li>
+        <li class="pb-3" id="phone">Phone:  ${response.phone_no}</li>
+        <li class="pb-3" id="positionLeft">Days left in the current position</li>
+        <li class="pb-3" id="premissionLeft">Days left to request permission</li>
+        <li class="pb-3" id="futureHoliday">Nearest future holiday</li>`
+
+    })
+}
+
+
+// Moment.js
+/*
+// current date and time
+var  m = moment();
+console.log(`toString() => ${m.toString()}`);
+console.log(`toISOString()=>${m.toISOString()}`);
+
+// create from iso 8601 string
+m=moment("2021-01-03T13:09:00.000-01:00");
+console.log(`toString() => ${m.toString()}`);
+console.log(`toISOString()=>${m.toISOString()}`);
+
+//parse using a format
+m=moment("14/06/2019 4:50PM", "DD/MM/YYYY h:mmA");
+
+console.log(`toString() => ${m.toString()}`);
+console.log(`toISOString()=>${m.toISOString()}`);
+
+
+//getter and setter
+const m=moment()
+console.log(m.toString());
+
+//getting units
+console.log(m.minutes());
+console.log(m.hour());
+console.log(m.week());
+console.log(m.year());
+console.log(m.get("quarter"));
+
+//setting units w/ moment.js
+m.minutes(65);
+m.hour(12);
+m.week(20);
+m.set("day", 9 )
+console.log(m.toString());
+
+
+//min & max of a particular set of moment objects
+// compare two moemnt objects
+const a=moment("01-01-2021")
+const n=moment("02-02-2021")
+console.log(moment.max(moment(), a).toString());
+console.log(moment.min(moment(), a).toString());
+
+
+//manupulate, add subtract
+const m = moment();
+
+console.log(`original moment: ${m.toString()}`);
+//add or subtract
+m.add(4,"hours");
+m.add(1,"h").add(12,"m");
+m.add({
+    "hours":1,
+    "minutes":20
+})
+m.subtract({
+    "hours":1,
+    "minutes":20
+})
+//return to the begining of the hour
+m.startOf("hour");
+m.startOf("year");
+m.endOf("day");
+
+console.log(`after manipulation: ${m.toISOString()}`);
+
+
+//format for display
+const m = moment();
+
+//console.log(m.format("[bfcxzrgso] dd mm yyyy"))
+m.locale("en-au")
+console.log(m.format("L"))
+
+
+
+//how long ago was this moment from now
+const m=moment("2019-05-13")
+console.log(m.fromNow())
+console.log(m.from("2019-05-15"))
+const a=moment().add(3,"days")
+console.log(a.calendar())
+const b=moment().add(30,"days")
+console.log(b.calendar())
+
+// find the diffeyfformrences between two moment objects
+const m=moment()
+const m1=moment("2020-05-13")
+console.log(m.diff(m1,"weeks"))
+
+const m1=moment("2020-05-13")
+const jsonObj={
+    momentObj:m1
+};
+console.log(JSON.stringify(jsonObj));
+
+*/
+
+
